explicit cast

+++ bn_mp_set_double.c
@@ -24,7 +24,7 @@ int mp_set_double(mp_int *a, double b)
    } cast;
    cast.dbl = b;
 
-   exp = (unsigned)(cast.bits >> 52) & 0x7FFU;
+   exp = (int)((unsigned)(cast.bits >> 52) & 0x7FFU);
    frac = (cast.bits & ((1ULL << 52) - 1ULL)) | (1ULL << 52);
 
    if (exp == 0x7FF) { /* +-inf, NaN */
